set x.y.z versions in Cargo.toml (#2204)
Set the x.y.z. versions in Cargo.toml files, exclusive of the submodule chunk_reads. Also raise and equalize versions in some cases. We are setting the x.y.z versions to reduce the likelihood
that merges between master and branches result in broken code whose resolution is complex.
The hope is that if the code will not compile after making a merge, then it will compile after
deleting the Cargo.lock file.