Add include directives to Cargo.tomls
These directives control which files are downloaded when someone gets
the crate from crates.io.  It is good practice to exclude test files,
READMEs, and so on in order to minimize the download.

The main impact of this is on the 50+mb of fasta files in vdj_ann, but
there's also a relatively large jpg image for pretty_trace.