initial minimal Rust GUI (#239)
Add an initial RUST GUI server/client, that works, but has several limitations, some of which need to be addressed
before anyone should use it.   See enclone_main/src/enclone_client.rs for the to do list.

In addition, there are many new crate dependencies, and for now, crate duplication.  We can try to fix some of the 
crate duplication after the next release of tokio.  The executable size is greatly increased.  We are accepting this 
tradeoff, with the hope that we can address it partially later.  Possibly the bulk of the change is due not to 
duplicated crates, but rather to addition of new crate dependencies, and we can't do much about that.

Release executable size increases from 46 to 134 MB, almost three fold.
Duplicated crates increase from 2 to 30.

Bloated executables are gross but it's not clear that it really matters that much.  

The changes do increase compilation time (which we do care about) but it seems OK given all we're getting in exchange.  
The CI tests appear greatly slowed down, but we can directly test essentially everything without using the CI.  The CI
tests would be faster with caching and if we could increase the number of CPUs.  On my Mac, cargo tb takes 4.5 minutes.

We add BSL-1.0 and OFL-1.1 to the list of allowed license types for crate dependencies, and manually accepted several 
other crate licenses.

Clock cycle speed test got 0.4% faster.  Go figure.  Peak memory increased by 0.7%.
Total test time appears to increase from 128 to 137 seconds.

Testing: ./test and enclone.test pass, and the GUI works.