delete five duplicate genes from the human VDJ reference (#116)
Delete the following genes from the human VDJ reference:
IGHV4-30-2, IGKV1D-33, IGKV1D-37, IGKV1D-39, IGKV2D-28.
These are identical to other genes in the reference, except that the reference provides a
longer 5'-UTR in one case. We observe that clonotypes having one of these genes often
have heterogeneous assignments between the gene and its duplicate, and that's bad, as it
implies that assignment of genes to clonotypes in these cases is effectively random.

Also 
cargo update -p regex
to get CI to pass.