Bump rust-toolbox (lockfile-only updates).
This updates the transitive dependency on pprof to 0.7.0 from 0.6.2.
The upper bound on pprof has increased further than that, but we can't
update it that far without ending up with other duplicated dependencies,
which is why this had to be done by hand rather than using dependabot.

This also updates hdf5 to allow a more recent version of parking_lot.