build: get prost-build out of critical path
Move generating of protobuf code into a separate crate so that
prost-build doesn't need to be compiled before we can compile
enclone_proto, and transitive dependents don't need it at all.
A unit test enforces that the file be up to date with whatever
the current version of prost-build creates.

This becomes especially important in newer versions of prost-build,
which depend on having protoc available externally in the environment;
this way, downstream crates do not need to have protoc installed.