Fix overflow bugs in VDJ FR3 start identification. (#316)
Consistently handle FR3 start window bounds in the face of sequences that are possibly shorter than the window we're searching in. Bail out early if the full search window does not overlap with the sequence, to ensure that the FR3 start computed based on a given contig is the same irrespective of the amount of prefix clipped in the contig.